+++ utils/runtime-fuzzer/src/generator/send_message.rs
@@ -16,12 +16,24 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
+use super::GenerationEnvironment;
 use gear_call_gen::{GearCall, SendMessageArgs};
 use gear_core::ids::ProgramId;
 use gear_utils::NonEmpty;
 use gear_wasm_gen::wasm_gen_arbitrary::{Result, Unstructured};
+use std::result::Result as StdResult;
 
 pub(crate) type SendMessageRuntimeData<'a> = (NonEmpty<&'a ProgramId>, u64);
+
+impl<'a> TryFrom<GenerationEnvironment<'a>> for SendMessageRuntimeData<'a> {
+    type Error = ();
+
+    fn try_from(env: GenerationEnvironment<'a>) -> StdResult<Self, Self::Error> {
+        let programs = NonEmpty::from_slice(&env.programs).ok_or(())?;
+
+        Ok((programs, env.max_gas))
+    }
+}
 
 pub(crate) fn generate(
     unstructured: &mut Unstructured,
